refactor: BlogCategories widget #308
