+++ Fanray.sln
@@ -58,7 +58,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sql", "sql", "{27E759ED-9E6
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ForkMeRibbon", "src\Plugins\ForkMeRibbon\ForkMeRibbon.csproj", "{22949558-4220-4879-8BB1-D03BCEFDD981}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogArchives", "src\Widgets\BlogArchives\BlogArchives.csproj", "{7BE27F20-9232-470F-ABC1-F0FCD54BD17D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlogArchives", "src\Widgets\BlogArchives\BlogArchives.csproj", "{7BE27F20-9232-470F-ABC1-F0FCD54BD17D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogCategories", "src\Widgets\BlogCategories\BlogCategories.csproj", "{84328837-CF2B-4AD3-A6FC-5D2D171996FA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -126,6 +128,10 @@ Global
 		{7BE27F20-9232-470F-ABC1-F0FCD54BD17D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7BE27F20-9232-470F-ABC1-F0FCD54BD17D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7BE27F20-9232-470F-ABC1-F0FCD54BD17D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{84328837-CF2B-4AD3-A6FC-5D2D171996FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{84328837-CF2B-4AD3-A6FC-5D2D171996FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{84328837-CF2B-4AD3-A6FC-5D2D171996FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{84328837-CF2B-4AD3-A6FC-5D2D171996FA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -150,6 +156,7 @@ Global
 		{153316E2-F584-4C7E-8090-F0D3E786378A} = {B073FE71-0D11-406D-8BE3-CC8CBBA831A9}
 		{22949558-4220-4879-8BB1-D03BCEFDD981} = {7B70F90E-8DFB-4215-9455-C9357D5C1B3A}
 		{7BE27F20-9232-470F-ABC1-F0FCD54BD17D} = {B073FE71-0D11-406D-8BE3-CC8CBBA831A9}
+		{84328837-CF2B-4AD3-A6FC-5D2D171996FA} = {B073FE71-0D11-406D-8BE3-CC8CBBA831A9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F1FC29A1-2BF1-46BD-9169-3C33C1C4494F}
