+++ src/Core/Fan.WebApp/Fan.WebApp.csproj
@@ -25,6 +25,7 @@
     <ProjectReference Include="..\..\Plugins\Shortcodes\Shortcodes.csproj" />
     <ProjectReference Include="..\..\Themes\Clarity\Clarity.csproj" />
     <ProjectReference Include="..\..\Widgets\BlogArchives\BlogArchives.csproj" />
+    <ProjectReference Include="..\..\Widgets\BlogCategories\BlogCategories.csproj" />
     <ProjectReference Include="..\..\Widgets\BlogTags\BlogTags.csproj" />
     <ProjectReference Include="..\Fan.Web\Fan.Web.csproj" />
   </ItemGroup>
