+++ src/Widgets/BlogCategories/Components/BlogCategories.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Fan.Blog.Models
-@using Fan.WebApp.Widgets.BlogCategories
+@using BlogCategories
 @model Tuple<IEnumerable<Category>, BlogCategoriesWidget>
 @{
     var categories = Model.Item1;
