+++ src/Widgets/BlogCategories/Manage/Widgets/BlogCategoriesEdit.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model BlogCategoriesEditModel
+@model BlogCategories.Manage.Widgets.BlogCategoriesEditModel
 @{
     ViewData["Title"] = "BlogCategories Widget Settings";
     Layout = "_SettingsLayout";
