+++ src/Widgets/BlogCategories/widget.json
